The brutalising of the app is done.
