--- mobile/build.gradle
@@ -1,55 +1,0 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
-
-android {
-    compileSdkVersion 29
-    defaultConfig {
-        applicationId "uk.co.ijhdev.trtlware"
-        minSdkVersion 22
-        targetSdkVersion 29
-        versionCode 6
-        versionName "1.0.2"
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-    compileOptions {
-        sourceCompatibility = 1.8
-        targetCompatibility = 1.8
-    }
-    buildToolsVersion = '28.0.3'
-}
-
-apply plugin: 'com.google.gms.google-services'
-apply plugin: 'io.fabric'
-
-dependencies {
-    wearApp project(path: ':wear')
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-    //google libs
-    implementation 'com.google.firebase:firebase-analytics:17.2.0'
-    implementation "com.google.android.gms:play-services-awareness:17.1.0"
-    implementation 'com.google.android.gms:play-services-wearable:17.0.0'
-    implementation 'com.google.code.gson:gson:2.8.5'
-    implementation 'com.google.firebase:firebase-perf:19.0.1'
-    implementation 'com.google.android.material:material:1.0.0'
-    //androidx
-    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    implementation 'androidx.appcompat:appcompat:1.1.0'
-    implementation 'androidx.recyclerview:recyclerview:1.0.0'
-    implementation 'androidx.cardview:cardview:1.0.0'
-    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
-    //retrofit
-    implementation 'com.squareup.retrofit2:retrofit:2.4.0'
-    implementation 'com.squareup.retrofit2:converter-gson:2.4.0'
-    //testing
-    androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0', {
-        exclude group: 'com.android.support', module: 'support-annotations'
-    })
-    testImplementation 'junit:junit:4.12'
-}
