--- mobile/src/main/AndroidManifest.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="uk.co.ijhdev.trtlware">
-
-    <!-- Required to act as a custom watch face. -->
-    <uses-permission android:name="android.permission.WAKE_LOCK" /> <!-- Required for complications to receive complication data and open the provider chooser. -->
-    <uses-permission android:name="com.google.android.wearable.permission.RECEIVE_COMPLICATION_DATA" />
-    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION"/>
-    <uses-permission android:name="android.permission.INTERNET" />
-
-    <application
-        android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
-        android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_launcher_round"
-        android:supportsRtl="true"
-        android:theme="@style/AppTheme">
-        <service android:enabled="true" android:name="uk.co.ijhdev.trtlware.Utils.WearableBackgroundListener" >
-                <intent-filter>
-                    <action android:name="com.google.android.gms.wearable.DATA_CHANGED" />
-                    <data android:scheme="wear" android:host="*"
-                        android:path="/start-activity" />
-                </intent-filter>
-            </service>
-        <meta-data
-            android:name="com.google.android.awareness.API_KEY"
-            android:value="AIzaSyBai1T96p7o62RejakjzNFpUe0t5rHe7lk"/>
-        <activity
-            android:name=".Activity.WareSettingsActivity"
-            android:label="@string/app_name"
-            android:screenOrientation="portrait">
-            <intent-filter>
-                <action android:name="android.intent.action.VIEW" />
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.DEFAULT" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-            <meta-data
-                android:name="com.google.android.gms.version"
-                android:value="@integer/google_play_services_version" />
-        </activity>
-        <service android:name="uk.co.ijhdev.trtlware.Utils.WearableBackgroundListener">
-            <intent-filter>
-                <action android:name="com.google.android.gms.wearable.DATA_CHANGED" />
-                <data android:host="*"
-                    android:pathPrefix="/phoneBattery"
-                    android:scheme="wear" />
-                <data android:host="*"
-                    android:pathPrefix="/weather"
-                    android:scheme="wear" />
-                <data android:host="*"
-                    android:pathPrefix="/price"
-                    android:scheme="wear" />
-            </intent-filter>
-        </service>
-    </application>
-
-</manifest>
