--- mobile/src/main/res/menu/tip_menu.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:id="@+id/btc_tip"
-        android:title="Tip BTC" />
-    <item
-        android:id="@+id/trtl_tip"
-        android:title="Tip TRTL" />
-
-</menu>
