--- mobile/src/main/res/values/colors.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <color name="colorPrimary">#00833d</color>
-    <color name="colorPrimaryDark">#212821</color>
-    <color name="colorAccent">#FFFFFF</color>
-</resources>
