+++ wear/src/main/java/uk/co/ijhdev/trtlware/Workers/PhoneBatteryWorker.kt
@@ -1,0 +1,18 @@
+package uk.co.ijhdev.trtlware.Workers
+
+import android.content.Context
+import android.os.BatteryManager
+import androidx.appcompat.app.AppCompatActivity
+
+/**
+ * Created by Seperot on 28/03/2018.
+ */
+
+class PhoneBatteryWorker{
+
+    fun getBatteryLevel(context: Context) : String {
+        val bm = context.getSystemService(AppCompatActivity.BATTERY_SERVICE) as BatteryManager
+        val batLevel = bm.getIntProperty(BatteryManager.BATTERY_PROPERTY_CAPACITY)
+        return "$batLevel %"
+    }
+}
