+++ wear/src/main/res/values/currency_array.xml
